Component Reusability for Pipelines
