+++ src/api/common/index.ts
@@ -180,7 +180,6 @@ export const fetchPipelines = async (perPage: number, page: number) => {
   };
 };
 
-
 export async function fetchResources(pipelineInstance: Pipeline) {
   const params = {
     limit: 100,
@@ -257,15 +256,5 @@ export async function fetchComputeInfo(
   return undefined;
 }
 
-export function hasCode(str: string) {
-  let hash = 0;
-  for (let i = 0; i < str.length; i++) {
-    hash = str.charCodeAt(i) + ((hash << 5) - hash);
-  }
-  return hash;
-}
 
-export function intToRGB(i: number) {
-  const c = (i & 0x00ffffff).toString(16).toUpperCase();
-  return "00000".substring(0, 6 - c.length) + c;
-}
+
