+++ src/components/feed/Pipelines/GeneralCompute.tsx
@@ -1,9 +1,8 @@
 import React, { useContext } from "react";
-import { List, Checkbox, Avatar } from "antd";
 import ChrisAPIClient from "../../../api/chrisapiclient";
 import { PipelineContext } from "../CreateFeed/context";
+import ListCompute from "./ListCompute";
 import { PipelineTypes } from "../CreateFeed/types/pipeline";
-import { hasCode, intToRGB } from "../CreateFeed/utils/pipelines";
 
 const GeneralCompute = ({
   currentPipelineId,
@@ -17,8 +16,7 @@ const GeneralCompute = ({
   const { state, dispatch } = useContext(PipelineContext);
   const [computes, setComputes] = React.useState<any[]>([]);
 
-  const generalCompute =
-    state.pipelineData[currentPipelineId].generalCompute;
+  const generalCompute = state.pipelineData[currentPipelineId].generalCompute;
 
   React.useEffect(() => {
     async function fetchCompute() {
@@ -27,12 +25,18 @@ const GeneralCompute = ({
         limit: 100,
         offset: 0,
       });
-      const computes = computeResourceList.getItems();
-      computes && setComputes(computes);
+      setComputes(computeResourceList.data);
     }
 
     fetchCompute();
   }, []);
+
+  const dispatchFn = (item: any) => {
+    dispatch({
+      type: PipelineTypes.SetGeneralCompute,
+      payload: { currentPipelineId, computeEnv: item.name },
+    });
+  };
 
   return (
     <div
@@ -42,48 +46,10 @@ const GeneralCompute = ({
       }}
       className="general-compute"
     >
-      <List
-        itemLayout="horizontal"
-        dataSource={computes ? computes : []}
-        renderItem={(item) => {
-          return (
-            <List.Item>
-              <List.Item.Meta
-                avatar={
-                  <>
-                    <Checkbox
-                      style={{
-                        marginRight: "0.5em",
-                      }}
-                      checked={
-                        generalCompute && generalCompute === item.data.name
-                          ? true
-                          : false
-                      }
-                      onClick={() => {
-                        dispatch({
-                          type: PipelineTypes.SetGeneralCompute,
-                          payload: {
-                            currentPipelineId,
-                            computeEnv: item.data.name,
-                          },
-                        });
-                      }}
-                    />
-
-                    <Avatar
-                      style={{
-                        background: `#${intToRGB(hasCode(item.data.name))}`,
-                      }}
-                    />
-                  </>
-                }
-                title={item.data.name}
-                description={item.data.description}
-              />
-            </List.Item>
-          );
-        }}
+      <ListCompute
+        computeList={computes}
+        generalCompute={generalCompute}
+        dispatchFn={dispatchFn}
       />
     </div>
   );
