+++ src/pages/DataLibrary/components/PACSLookup/QueryBuilder.tsx
@@ -40,8 +40,6 @@ export const QueryBuilder: React.FC<QueryBuilderProps> = ({
     type: PFDCMQueryTypes.PMRN,
   } as PFDCMQuery);
 
-  console.log("PACSSERVICES", PACSservices);
-
   const setQueryType = (type: PFDCMQueryTypes) => {
     setQuery({ ...query, type } as PFDCMQuery);
   };
